# activity/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Activity Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/activity-groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Lists activity groups with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity Groups"
                ],
                "summary": "List activity groups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by creator ID",
                        "name": "creator_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (draft, active, cancelled)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by title (partial match)",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit number of results (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Groups retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_ActivityGroupListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Creates a new activity group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity Groups"
                ],
                "summary": "Create an activity group",
                "parameters": [
                    {
                        "description": "Activity group data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.CreateActivityGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Group created successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_CreateActivityGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/activity-groups/discover": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Discovers public activity groups with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity Groups"
                ],
                "summary": "Discover activity groups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by city",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by country",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by activity type",
                        "name": "activity_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit number of results (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Groups retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_ActivityGroupListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
        "/api/v1/activity-groups/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Retrieves an activity group by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity Groups"
                ],
                "summary": "Get an activity group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_ActivityGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Group not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Updates an existing activity group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity Groups"
                ],
                "summary": "Update an activity group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.UpdateActivityGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group updated successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Group not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Cancels (soft-deletes) an activity group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity Groups"
                ],
                "summary": "Cancel an activity group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group cancelled successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Group not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/api/v1/activity-groups/{id}/activate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": [],
                        "GatewayKeyAuth": []
                    }
                ],
                "description": "Activates a draft activity group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activity Groups"
                ],
                "summary": "Activate an activity group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Activity Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group activated successfully",
                        "schema": {
                            "$ref": "#/definitions/api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Group not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Response-any": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success"
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
        "api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_ActivityGroupListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.ActivityGroupListResponse"
                        }
                    ]
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
        "api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_ActivityGroupResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.ActivityGroupResponse"
                        }
                    ]
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
...
        "api.Response-github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos_CreateActivityGroupResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "only on success",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.CreateActivityGroupResponse"
                        }
                    ]
                },
                "error": {
                    "description": "only on error",
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.APIError"
                        }
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                }
            }
        },
...
        "github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.ActivityGroupListResponse": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.ActivityGroupResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_rasparac_rekreativko-api_activity_internal_interfaces_http_dtos.CreateActivityGroupRequest": {
            "type": "object",
            "required": [
                "activity_type",
                "difficulty_level",
                "location_city",
                "location_country",
                "timezone",
                "title",
                "visibility"
            ],
            "properties": {
                "activity_type": {
                    "type": "string",
                    "enum": [
                        "hiking",
                        "cycling",
                        "running",
                        "swimming",
                        "yoga",
                        "gym",
                        "other"
                    ],
                    "example": "running"
                },
                "default_capacity": {
                    "type": "integer",
                    "maximum": 1000,
                    "minimum": 1,
                    "example": 30
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "A group for early morning runners in Belgrade"
                },
                "difficulty_level": {
                    "type": "string",
                    "enum": [
                        "beginner",
                        "intermediate",
                        "advanced"
                    ],
                    "example": "beginner"
                },
                "location_city": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Belgrade"
                },
                "location_country": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Serbia"
                },
                "timezone": {
                    "type": "string",
                    "example": "Europe/Belgrade"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Morning Running Group"
                },
                "visibility": {
                    "type": "string",
                    "enum": [
                        "public",
                        "private"
                    ],
                    "example": "public"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GatewayKeyAuth": {
            "description": "Key added automatically by the gateway when proxying requests; required directly here only when bypassing the gateway.",
            "type": "apiKey",
            "name": "X-Gateway-Key",
            "in": "header"
        }
    },
    "security": [
        {
            "GatewayKeyAuth": []
        }
    ]
}